requirements.txt
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
click==7.1.2
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.7
django-crispy-forms==1.8.1
django-filter==2.3.0
djangorestframework==3.11.0
entrypoints==0.3
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numpy==1.18.4
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.8.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.5.14
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
